@chris-allan 2015-04-23
Integration tests to cover the cases fixed by 230a2be473c92379a72c871af2a445dcf5e8846e now included on `chris-allan/openmicroscopy:optimize-ofp-develop`. Should pull in cleanly and address the scenarios that were noticed by @aleksandra-tarkowska on #3695.