@sbesson 2015-04-24
Retested with a 3K files dataset. Works as expected. Given the state of the integration tests https://ci.openmicroscopy.org/job/OMERO-5.1-merge-integration-python/lastCompletedBuild/testReport/test.integration.gatewaytest.test_fs/TestFileset/, merging this PR for 5.1.1. Thanks all.